requirements.txt
google-cloud-bigquery
google-cloud-storage
dotenv
pandas
pyarrow
openpyxl
selenium
webdriver-manager
beautifulsoup4
python-dotenv